refactor: 残余 String ID 代码异味清理 — CatalogService/JWT/Audit/DataTable/Archive
- CatalogService: HashMap<String, T> → HashMap<i64, T>，移除 30+ 处 .to_string()
- JWT CurrentUser: id/role_id String → i64，Claims 解析边界集中在 From<Claims>
- AuditEntry: operator_id Option<String> → Option<i64>，SQLite 存储层做 String 转换
- DataTable: getRowKey 返回类型 string → string | number，移除 7 处 String() 包装
- Archive API: SplitItemDetail 补 instance_id 字段，ArchivedEvent.event_id string → number

+++ edge-server/src/db/repository/system_state.rs
@@ -73,10 +73,10 @@ pub async fn init_genesis(pool: &SqlitePool, genesis_hash: String) -> RepoResult
 }
 
 /// Atomically increment order_count and return the new value
-pub async fn get_next_order_number(pool: &SqlitePool) -> RepoResult<i32> {
+pub async fn get_next_order_number(pool: &SqlitePool) -> RepoResult<i64> {
     let now = shared::util::now_millis();
     let new_count = sqlx::query_scalar!(
-        r#"UPDATE system_state SET order_count = order_count + 1, updated_at = ?1 WHERE id = ?2 RETURNING order_count as "order_count!: i32""#,
+        r#"UPDATE system_state SET order_count = order_count + 1, updated_at = ?1 WHERE id = ?2 RETURNING order_count as "order_count!: i64""#,
         now,
         SINGLETON_ID
     )
